requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
plotly
altair
fpdf
Pillow